requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
PyJWT==2.8.0
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
gunicorn==21.2.0
pydantic==2.11.7
asyncpg==0.28.0  # 添加异步PostgreSQL支持
sqlalchemy-utils==0.41.1  # 添加SQLAlchemy工具
python-dotenv==1.0.0  # 确保环境变量加载
